LayoutTests/fast/css/font-family-trailing-bracket-gunk.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<span id="testspan" style="font-family: 'Arial [ding dong]', 'Helvetica [Xft]', Courier">foo</span>
<script>

description("This test checks that we don't strip trailing gunk in brackets from font families.");

var spanElement = document.getElementById('testspan');

var computedStyle = window.getComputedStyle(spanElement);
shouldBe("spanElement.innerHTML", "'foo'");
shouldBe("computedStyle.getPropertyValue('font-family')", "\"'Arial [ding dong]', 'Helvetica [Xft]', Courier\"");

document.body.removeChild(spanElement);

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
