third_party/blink/web_tests/platform/generic/fast/css/font-family-trailing-bracket-gunk-expected.txt

This test checks that we don't strip trailing gunk in brackets from font families.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spanElement.innerHTML is 'foo'
PASS computedStyle.getPropertyValue('font-family') is '"Arial [ding dong]", "Helvetica [Xft]", Courier'
PASS successfullyParsed is true

TEST COMPLETE
